# apps/web/src/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/export": {
      "get": {
        "description": "Same sheets and columns as the import: exporting then importing gives back the same data.",
        "operationId": "ExportController_exportSpreadsheet",
        "parameters": [
          {
            "name": "format",
            "required": true,
            "in": "query",
            "schema": {
              "example": "xlsx",
              "type": "string",
              "enum": [
                "xlsx",
                "ods"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "application/vnd.oasis.opendocument.spreadsheet": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_FAILED",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Export the applications and skills as a spreadsheet (.xlsx or .ods)",
        "tags": [
          "export"
        ]
      }
    }
  },
  "info": {
    "title": "OpenJobSeekR API",
    "description": "Local job application tracker. Errors are RFC 9457 problem details (application/problem+json) with a stable `code`.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "FieldErrorDto": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "example": "company"
          },
          "constraints": {
            "example": [
              "isNotEmpty"
            ],
            "description": "Failed constraint names, sorted",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "field",
          "constraints"
        ]
      },
      "ProblemDetailsDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "urn:openjobseekr:error:application-not-found"
          },
          "title": {
            "type": "string",
            "example": "Application not found"
          },
          "status": {
            "type": "number",
            "example": 404
          },
          "detail": {
            "type": "string",
            "description": "Developer message about this occurrence"
          },
          "instance": {
            "type": "string",
            "example": "/applications/6c3f4d2e-0000-4000-8000-000000000001"
          },
          "code": {
            "type": "string",
            "enum": [
              "VALIDATION_FAILED",
              "BAD_REQUEST",
              "UNAUTHORIZED",
              "INVALID_CREDENTIALS",
              "FORBIDDEN",
              "NOT_FOUND",
              "APPLICATION_NOT_FOUND",
              "SKILL_NOT_FOUND",
              "CONFLICT",
              "EMAIL_ALREADY_USED",
              "SKILL_NAME_ALREADY_USED",
              "PAYLOAD_TOO_LARGE",
              "IMPORT_UNSUPPORTED_FILE",
              "IMPORT_INVALID_STRUCTURE",
              "IMPORT_INVALID_DATA",
              "IMPORT_TOO_MANY_ROWS",
              "TOO_MANY_REQUESTS",
              "INTERNAL_ERROR",
              "SERVICE_UNAVAILABLE"
            ],
            "description": "Stable code, translated by the web app"
          },
          "errors": {
            "description": "Invalid fields (VALIDATION_FAILED), or spreadsheet cells (IMPORT_*)",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldErrorDto"
            }
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "instance",
          "code"
        ]
      }
    }
  }
}